+++ lib/mws/shipment_invoicing.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'mws/shipment_invoicing/client'